procurement_purchase_no_grouping_origin/models/res_company.py
from odoo import fields, models

class ResCompany(models.Model):
    _inherit = "res.company"

    procured_purchase_grouping = fields.Selection(
        selection_add=[
            ("origin", "Group by Origin"),
        ],
        ondelete={"origin": "set default"},
    )